requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose
moviepy
pydub
aiofiles
Pillow
opencv-python
mediapipe